+++ aiokraken/rest/schemas/kabtype.py
