--- requirements.txt
@@ -1,3 +1,0 @@
-aiohttp
-aiodns
-cchardet
